Added basic led subsystem
Added constants which still need to be updated to real values
added a configure feedback method in robot container. Nothing is in the method currently, but it will be used for triggers for led color commands in the future. also added red led as default command, and a rainbow command which is unused